rename NSTimer convenient method
`scheduledTimerWithTimeInterval(:repeats:handler:)` is sherlocked by
iOS 10/macOS 11.12/watchOS 3/tvOS 10 SDK. In detail, Apple added a class method
on NSTimer that has the same signature except the label for the last parameter:

    scheduledTimerWithTimeInterval(repeats:block:)

Call sites that uses trailing closure syntax will not compile in Xcode 10 due
to the ambiguity here.

For projects that need to support OS versions older than the ones listed above,
they need a way to keep using the ClosureKit version. Renaming the label on
the second argument is a solution (suggested by @keith and @mallorypaine).